# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mixtli API",
    "version": "1.0.0",
    "description": "Server-upload API for Mixtli (Cloudflare R2)."
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Healthcheck",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness probe",
        "responses": {
          "200": {
            "description": "Ready"
          },
          "503": {
            "description": "Not ready"
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "summary": "Upload a file via server",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "downloadUrl": {
                      "type": "string"
                    },
                    "publicUrl": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/signget": {
      "get": {
        "summary": "Sign a temporary GET URL for an existing key",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 600
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    }
  }
}